Extract DioHttpClient into dio_http_client.dart
